move service screen to view-component
