+++ bridge/client/app/_views/ktb-service-view/ktb-service-view.component.html
@@ -1,0 +1,91 @@
+<div class="container" fxFlex="34" fxLayout="column" fxLayoutGap="15px" *ngIf="project$ | async as project">
+  <div fxFlex fxLayout="column" fxLayoutGap="15px">
+    <div>
+      <dt-info-group>
+        <dt-info-group-title>
+          <p class="mb-0"><span [textContent]="project.getServices().length"></span> Services</p>
+        </dt-info-group-title>
+      </dt-info-group>
+    </div>
+    <div fxLayout="row" fxLayoutAlign="start center" *ngIf="project.getServices().length == 0">
+      <dt-icon class="icon" name="information"></dt-icon>
+      <p>No service onboarded yet. Follow the intructions to <a [href]="'/manage/service/#onboard-a-service' | keptnUrl" target="_blank" rel="noopener noreferrer">onboard a service</a>.</p>
+    </div>
+    <ktb-expandable-tile *ngFor="let service of project.getServices()" [expanded]="service.serviceName == serviceName">
+      <ktb-expandable-tile-header>
+        <dt-info-group>
+          <dt-info-group-title>
+            <div fxLayout="row">
+              <div fxFlex>
+                <h2 class="m-0" [textContent]="service.serviceName"></h2>
+              </div>
+              <div fxLayout="column" fxLayoutAlign="start end" *ngIf="eventTypes">
+                <div class="buttons-container" fxLayout="row" fxLayoutAlign="start start">
+                  <button dt-icon-button (click)="$event.stopPropagation()" [dtContextDialogTrigger]="filterEventsDialog" variant="nested" aria-label="Open context dialog">
+                    <dt-icon name="filter"></dt-icon>
+                  </button>
+                  <dt-context-dialog #filterEventsDialog aria-label="Open context dialog" ariaLabelClose="Close context dialog">
+                    <p *ngFor="let eventType of eventTypes">
+                      <dt-checkbox (change)="filterEvents($event, eventType)" [checked]="true || isFilteredEvent(eventType)" [textContent]="eventType"></dt-checkbox>
+                    </p>
+                  </dt-context-dialog>
+                </div>
+              </div>
+            </div>
+          </dt-info-group-title>
+          <div *ngIf="service.roots">
+            <p class="m-0 mb-1 mt-1" *ngIf="project.getLatestArtifact(service) as artifact; else noDeployment">
+              <span class="bold">Last processed artifact: </span><span *ngIf="artifact.getShortImageName()" [textContent]="artifact.getShortImageName()"></span><span *ngIf="!artifact.getShortImageName()">unknown</span>
+            </p>
+          </div>
+          <button dt-button disabled variant="nested" *ngIf="!service.roots">
+            <dt-loading-spinner></dt-loading-spinner>
+            Loading ...
+          </button>
+        </dt-info-group>
+      </ktb-expandable-tile-header>
+      <div class="mb-1" fxLayout="row" fxLayoutAlign="end center">
+        <dt-tag-list aria-label="fetch-info">
+          <dt-tag>Last time fetched: <span [textContent]="getRootsLastUpdated(project) | amCalendar:dateUtil.getCalendarFormats(true)"></span></dt-tag>
+        </dt-tag-list>
+      </div>
+      <ktb-root-events-list [events]="getFilteredRoots(service.roots)" [selectedEvent]="currentRoot" (selectedEventChange)="selectRoot($event)"></ktb-root-events-list>
+      <button dt-button disabled variant="nested" *ngIf="!service.roots">
+        <dt-loading-spinner></dt-loading-spinner>
+        Loading ...
+      </button>
+    </ktb-expandable-tile>
+    <div class="mb-3"></div>
+  </div>
+</div>
+<div class="container dark" fxFlex="66" fxLayout="column" fxLayoutGap="15px">
+  <div *ngIf="currentRoot" fxFlex fxLayout="column" fxLayoutGap="15px">
+    <div>
+      <dt-info-group>
+        <dt-info-group-title>
+          <div fxFlex fxLayout="row" fxLayoutAlign="flex-start center" fxLayoutGap="15px">
+            <div fxLayout="row" fxLayoutAlign="start center">
+              <button dt-icon-button disabled variant="nested" *ngIf="currentRoot.isFaulty()"><dt-icon name="criticalevent"></dt-icon></button>
+              <p *ngIf="currentRoot.getShortImageName(); else noImageName" [textContent]="currentRoot.getShortImageName()"></p>
+              <ng-template #noImageName>
+                <p class="mr-1" *ngIf="currentRoot.getService()" [textContent]="currentRoot.getService()"></p>
+              </ng-template>
+            </div>
+          </div>
+        </dt-info-group-title>
+        <p class="m-0" *ngIf="currentRoot.data.configurationChange?.values"><span class="bold">New artifact: </span><span [textContent]="currentRoot.data.configurationChange.values.image"></span></p>
+        <p class="m-0"><span class="bold">Keptn ID: </span><span [textContent]="currentRoot.shkeptncontext"></span></p>
+      </dt-info-group>
+    </div>
+    <ktb-events-list [events]="currentRoot.traces" [focusedEventId]="eventId"></ktb-events-list>
+    <div class="mb-3" fxLayout="row" fxLayoutAlign="end center">
+      <dt-tag-list aria-label="fetch-info">
+        <dt-tag>Last time fetched: <span [textContent]="getTracesLastUpdated(currentRoot) | amCalendar:dateUtil.getCalendarFormats(true)"></span></dt-tag>
+        <dt-tag *ngIf="showReloadButton(currentRoot)"><button dt-button (click)="loadTraces(currentRoot)"><dt-icon name="refresh"></dt-icon> Reload</button></dt-tag>
+      </dt-tag-list>
+    </div>
+  </div>
+</div>
+<ng-template #noDeployment>
+  <p class="m-0">Service not deployed yet. Use <a [href]="'/reference/cli/commands/keptn_send_event_new-artifact/' | keptnUrl" target="_blank" rel="noopener noreferrer">keptn send event new-artifact</a> to trigger a deployment.</p>
+</ng-template>
