+++ bridge/client/app/_views/ktb-service-view/ktb-service-view.component.scss
@@ -1,0 +1,2 @@
+@import '../../../variables';
+@import '~@dynatrace/barista-components/core/src/style/variables';
