+++ bridge/client/app/project-board/project-board.component.ts
@@ -1,20 +1,14 @@
 import {ChangeDetectorRef, Component, OnDestroy, OnInit} from '@angular/core';
-import {filter, map, startWith, switchMap, take, takeUntil} from "rxjs/operators";
-import {Observable, Subject, Subscription, timer} from "rxjs";
+import {map, takeUntil} from "rxjs/operators";
+import {Observable, Subject} from "rxjs";
 import {ActivatedRoute, Router} from "@angular/router";
-import {Location} from "@angular/common";
 
-import * as moment from 'moment';
-
-import {Root} from "../_models/root";
 import {Project} from "../_models/project";
+import {Trace} from "../_models/trace";
 
 import {DataService} from "../_services/data.service";
 import {ApiService} from "../_services/api.service";
 import {DateUtil} from "../_utils/date.utils";
-import {Trace} from "../_models/trace";
-import {DtCheckboxChange} from "@dynatrace/barista-components/checkbox";
-import {EVENT_LABELS} from "../_models/event-labels";
 
 @Component({
   selector: 'app-project-board',
@@ -26,33 +20,19 @@ export class ProjectBoardComponent implements OnInit, OnDestroy {
   private readonly unsubscribe$ = new Subject<void>();
 
   public project$: Observable<Project>;
-
-  public currentRoot: Root;
-  public error: string = null;
-
-  private _rootEventsTimerInterval = 30;
-  private _tracesTimerInterval = 10;
-  private _tracesTimer: Subscription = Subscription.EMPTY;
-
-  public projectName: string;
-  public serviceName: string;
   public contextId: string;
-  public eventId: string;
 
+  public error: string = null;
   public view: string = 'services';
 
-  public eventTypes: string[] = [];
-  public filterEventTypes: string[] = [];
-
-  constructor(private _changeDetectorRef: ChangeDetectorRef, private router: Router, private location: Location, private route: ActivatedRoute, private dataService: DataService, private apiService: ApiService, public dateUtil: DateUtil) { }
+  constructor(private _changeDetectorRef: ChangeDetectorRef, private router: Router, private route: ActivatedRoute, private dataService: DataService, private apiService: ApiService) { }
 
   ngOnInit() {
     this.route.params
       .pipe(takeUntil(this.unsubscribe$))
       .subscribe(params => {
         if(params["shkeptncontext"]) {
-          this.contextId = params["shkeptncontext"];
-          this.apiService.getTraces(this.contextId)
+          this.apiService.getTraces(params["shkeptncontext"])
             .pipe(
               map(response => response.body),
               map(result => result.events||[]),
@@ -79,15 +59,8 @@ export class ProjectBoardComponent implements OnInit, OnDestroy {
               }
             });
         } else {
-          this.projectName = params["projectName"];
-          this.serviceName = params["serviceName"];
           this.contextId = params["contextId"];
-          this.eventId = params["eventId"];
-          this.currentRoot = null;
-          this.filterEventTypes = [];
-
           this.project$ = this.dataService.getProject(params['projectName']);
-
           this.project$
             .pipe(takeUntil(this.unsubscribe$))
             .subscribe(project => {
@@ -97,84 +70,8 @@ export class ProjectBoardComponent implements OnInit, OnDestroy {
             }, error => {
               this.error = 'projects';
             });
-
-          this.dataService.roots
-            .pipe(takeUntil(this.unsubscribe$))
-            .subscribe(roots => {
-              if(roots) {
-                if(!this.currentRoot)
-                  this.currentRoot = roots.find(r => r.shkeptncontext == params["contextId"]);
-                this.eventTypes = this.eventTypes.concat(roots.map(r => r.type)).filter((r, i, a) => a.indexOf(r) === i);
-              }
-              if(this.currentRoot && !this.eventId)
-                this.eventId = this.currentRoot.traces[this.currentRoot.traces.length-1].id;
-            });
-
-          timer(0, this._rootEventsTimerInterval*1000)
-            .pipe(
-              startWith(0),
-              switchMap(() => this.project$),
-              filter(project => !!project && !!project.getServices())
-            )
-            .pipe(takeUntil(this.unsubscribe$))
-            .subscribe(project => {
-              this.dataService.loadServices(project);
-              this.dataService.loadRoots(project);
-            });
         }
       });
-  }
-
-  selectRoot(event: any): void {
-    this.projectName = event.root.getProject();
-    this.serviceName = event.root.getService();
-    this.contextId = event.root.data.shkeptncontext;
-    this.eventId = null;
-    if(event.stage) {
-      let focusEvent = event.root.traces.find(trace => trace.data.stage == event.stage);
-      let routeUrl = this.router.createUrlTree(['/project', focusEvent.getProject(), focusEvent.getService(), focusEvent.shkeptncontext, focusEvent.id]);
-      this.eventId = focusEvent.id;
-      this.location.go(routeUrl.toString());
-    } else {
-      let routeUrl = this.router.createUrlTree(['/project', event.root.getProject(), event.root.getService(), event.root.shkeptncontext]);
-      this.eventId = event.root.traces[event.root.traces.length-1].id;
-      this.location.go(routeUrl.toString());
-    }
-
-    this.currentRoot = event.root;
-    this.loadTraces(this.currentRoot);
-  }
-
-  selectDeployment(deployment: Trace, project: Project) {
-    this.selectRoot({
-      root: project.getServices().find(service => service.serviceName === deployment.data.service).roots.find(root => root.shkeptncontext === deployment.shkeptncontext),
-      stage: deployment.data.stage
-    });
-  }
-
-  loadTraces(root: Root): void {
-    this._tracesTimer.unsubscribe();
-    if(moment().subtract(1, 'day').isBefore(root.time)) {
-      this._tracesTimer = timer(0, this._tracesTimerInterval*1000)
-        .subscribe(() => {
-          this.dataService.loadTraces(root);
-        });
-    } else {
-      this.dataService.loadTraces(root);
-      this._tracesTimer = Subscription.EMPTY;
-    }
-  }
-
-  getRootsLastUpdated(project: Project): Date {
-    return this.dataService.getRootsLastUpdated(project);
-  }
-
-  getTracesLastUpdated(root: Root): Date {
-    return this.dataService.getTracesLastUpdated(root);
-  }
-
-  showReloadButton(root: Root) {
-    return moment().subtract(1, 'day').isAfter(root.time);
   }
 
   loadProjects() {
@@ -185,31 +82,8 @@ export class ProjectBoardComponent implements OnInit, OnDestroy {
     this.view = view;
   }
 
-  filterEvents(event: DtCheckboxChange<string>, eventType: string): void {
-    let index = this.filterEventTypes.indexOf(eventType);
-    if(index == -1) {
-      this.filterEventTypes.push(eventType);
-    } else {
-      this.filterEventTypes.splice(index, 1);
-    }
-  }
-
-  isFilteredEvent(eventType: string) {
-    return this.filterEventTypes.indexOf(eventType) == -1;
-  }
-
-  getEventLabel(key): string {
-    return EVENT_LABELS[key] || key;
-  }
-
-  getFilteredRoots(roots: Root[]) {
-    if(roots)
-      return roots.filter(r => this.filterEventTypes.indexOf(r.type) == -1);
-  }
-
   ngOnDestroy(): void {
     this.unsubscribe$.next();
-    this._tracesTimer.unsubscribe();
   }
 
 }
